Refactor to fix build system / fix readme
