+++ README.md
@@ -4,8 +4,8 @@ YSO standard classification
 This Python package is used to classify young stellar objects using the
 standard classification scheme originally devised by Lada et al. (1984).
 
-It determines the infrared spectral index \(\alpha_{IR}\) in the range
-from \(2\,\mu\mathrm{m}\) to \(20\,\mu\mathrm{m}\), estimating the slope
+It determines the infrared spectral index $$\alpha_{IR}$$ in the range
+from $$2\,\mu\mathrm{m}$$ to $$20\,\mu\mathrm{m}$$, estimating the slope
 of the spectral energy distribution. Based on the measured slope, the
 YSOs are divided into 5 distinct classes (0/I, flat spectrum, II,
 III with debris disk, III no disk/MS star).
