+++ src/standard_classification/__init__.py
@@ -1,0 +1,2 @@
+from . import cls
+from ._version import __version__
