+++ src/standard_classification/_version.py
