+++ src/standard_classification/extinction/__init__.py
@@ -1,0 +1,1 @@
+#from . import fun #extinction, get_src_ext, query_ext
