+++ src/standard_classification/extinction/bla.py
@@ -1,0 +1,2 @@
+def fun():
+    print("fun")
