+++ src/standard_classification/extinction/extinction.py
