--- standard_classification/__init__.py
@@ -1,2 +1,0 @@
-from standard_classification import classify
-from ._version import __version__
