--- standard_classification/extinction/__init__.py
@@ -1,1 +1,0 @@
-from . import extinction, get_src_ext, query_ext
